Add file "evaluators.py"
In this module the function "evaluate_anomaly_detector()" is defined.
This function applies an instance af the "MinMaxAnomalyDetector" to a pandas.DataFrame
and checks if the data points marked as anomalous are detected by the detector.
The output of the function is a dict: the keys are the labels of the inserted anomalies
and the elements are pandas.Series, containing the timestamp of the anomalous value and one boolean
for each variable of the DataFrame. The bool is True if the anomaly has been detected for that variable, False otherwise.