Fix the evaluation strategy
Now the function "evaluate_anomaly_detector()" looks for the rows of the input dataset both marked and detected as anomalous.
The return value of the function is still a dict. The keys of the dict are the labels of the synthetic anomalies inside
the timeseries; the value associated with the key changes if the anomaly has been detected or not:
Key: detected anomaly --> Value: pandas.DataFrame with the rows of the timeseries containing the detected anomalous data points.
Key: non detected anomaly --> Value: False