Manage zero divisions
About the evaluation on a single series:
Before, the calculation of the 'anomalies_ratio' value was a zero division for not anomalous series.
Now, 'anomalies_ratio' value is "None" for not anomalous series.
About the evaluation on a dataset:
Now, the value 'anomalies_ratio' is calculated averaging only on anomalous series