Fix bugs
About "_point_eval_with_events_strategy()"
- The event evaluation was not taking into account not detected anomalies. The "breakdown_info" dictionary had only
detected anomalies as keys; the for cycle for calculating the anomaly ratio (anomaly_counts/inserted_anomalies)
was done on keys of "inserted_events_by_type", so a key error was raised for not detected anomalies.

- The dataframe was filtrated using the column name "_anomaly" to get the model flags but the "apply()" method of the model
adds the column "anomaly". Probably this is the reason why the evaluation on models like NHARAnomalyDetector was giving
0.0 everywhere.