Delete the instance of "PeriodicAverageAnomalyDetector()" in "test_correct_counting_false_positives_with_events_strategy()"
This test checks if true positives and false positives are correctly counted using a dataset with only 1 series.
In this case, the ratio between true positives (false positives) count and true positives rate (false positives ratio)
must give the total number of inserted anomalies (lenght of the series).
Tests pass for "MinMaxAnomalyDetector()" and "NHARAnomalyDetector()" but not for the "PeriodicAverageAnomalyDetector()"
because the output of the model is a shorter series than the input one.