Fix the problem with the periodic average model
Now the false positives ratio is computed as the ratio between false positives counts and the length
of the column 'anomaly_labels', which is equal to the length of the series of the dataset.
This was done because the function "_point_eval_with_events_strategy()" takes in input the flagged series
(output of the "apply()" method of the model) that for the periodic average model has a shorter length than the
original series of the dataset.